requirements.txt
weave
openai
python-dotenv
pinecone
numpy
scikit-learn
weave[scorers]
llama-cloud-services
streamlit
pydantic